Added basic redirect functionality
Agent can redirect by sending /REDIRECT through the chat.

Chat now checks for the redirect command on the client side. If so,
the current conversation is closed and disconnected, and a new agent
is requested via the new route /chat/redirect (since the user does not
require a new guest token for the redirect).

Issues:
- Currently, the matchmaker takes care of the redirect. May be good to
  let the agent decide who to redirect to.
- Would also be good to display some loading message to the user while
  processing the redirect.